move onboarding step machine into config/credentials.ts
